[REF] tg_pos: merged TestPosCustomerButtonVisible and TestPosCustomerButtonHidden into TestPosUi
