+++ tg_pos/tests/__init__.py
@@ -1,2 +1,1 @@
-from . import test_pos_button_hidden
-from . import test_pos_button_visible
+from . import test_pos_ui
